requirements.txt
discord.py==2.3.2
httpx==0.27.0
Pillow==10.3.0
hearthstone==1.2.0
python-dotenv==1.0.1
